feat: refactor página de configuración de usuario con nuevos componentes y validaciones
